requirements.txt
pandas>=2.0
numpy>=1.24
matplotlib>=3.7
seaborn>=0.12
scikit-learn>=1.3
openpyxl>=3.1
jupyter>=1.0
python-docx>=1.1
pdfplumber>=0.10
py7zr>=0.20